+++ CHANGELOG.md
@@ -91,6 +91,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Replace deprecated wasm features [#1988](https://github.com/provenance-io/provenance/pull/1988).
 * Add `UpdateParams` and `Params` query rpc endpoints to modules.
   * `ibcratelimit` add `UpdateParams` endpoint and deprecate `GovUpdateParams` [#1984](https://github.com/provenance-io/provenance/pull/1984).
+  * `attribute` add `UpdateParams` endpoint and cli [#1987](https://github.com/provenance-io/provenance/pull/1987).
 
 ### Client Breaking